% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-vector.R
\name{as_sparse_double}
\alias{as_sparse_double}
\title{Coerce numeric vector to sparse double}
\usage{
as_sparse_double(x, default = 0)
}
\arguments{
\item{x}{a numeric vector.}

\item{default}{default value to use. Defaults to \code{0}.}
}
\description{
Coerce numeric vector to sparse double
}
\examples{
x_dense <- c(3, 0, 2, 0, 0, 0, 4, 0, 0, 0)
x_sparse <- as_sparse_double(x_dense)
x_sparse

is_sparse_double(x_sparse) 
}
