% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-predicates.R
\name{is_sparse_vector}
\alias{is_sparse_vector}
\title{Sparse vector type checkers}
\usage{
is_sparse_vector(x)
}
\arguments{
\item{x}{value to be checked.}
}
\description{
Sparse vector type checkers
}
\examples{
x_sparse <- sparse_double(c(pi, 5, 0.1), c(2, 5, 10), 10)
x_dense <- c(0, pi, 0, 0, 0.5, 0, 0, 0, 0, 0.1)

is_sparse_vector(x_sparse)
is_sparse_vector(x_dense)

# Forced materialization
is_sparse_vector(x_sparse[])
}
