% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{new_sparse_vector}
\alias{new_sparse_vector}
\title{Create new sparse vector}
\usage{
new_sparse_vector(values, positions, length)
}
\arguments{
\item{values}{Vector of values}

\item{positions}{Vector of positions, must be integers and same length as
values.}

\item{length}{Length of resulting vector}
}
\value{
A sparse_vector object.
}
\description{
Create new sparse vector
}
\examples{
new_sparse_vector(1, 4, 10)
}
