% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{sparse_double}
\alias{sparse_double}
\title{Create sparse double vector}
\usage{
sparse_double(value, position, length)
}
\arguments{
\item{value}{Numeric vector, values of non-zero entries.}

\item{position}{integer vector, indices of non-zero entries.}

\item{length}{Integer, Length of vector.}
}
\description{
Create sparse double vector
}
\details{
For developers:

setting \code{options("sparsevctrs.verbose_materialize" = TRUE)} will print a
message each time a sparse vector has been forced to materialize.
}
\examples{
sparse_double(numeric(), integer(), 10)

sparse_double(c(pi, 5, 0.1), c(2, 5, 10), 10)

str(
  sparse_double(c(pi, 5, 0.1), c(2, 5, 10), 1000000000)
)
}
