% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{sparse_to_tibble}
\alias{sparse_to_tibble}
\title{Convert sparse matrices to tibbles}
\usage{
sparse_to_tibble(x)
}
\arguments{
\item{x}{a sparse matrix}
}
\value{
a tibble
}
\description{
Convert sparse matrices to tibbles
}
\examples{
\dontshow{if (rlang::is_installed("rsparse")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# data("movielens100k", package = "rsparse")

# sparse_to_tibble(movielens100k)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=tibble_to_sparse]{tibble_to_sparse()}}
}
