% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{sparse_vector}
\alias{sparse_vector}
\title{Create new sparse vector}
\usage{
sparse_vector(values = double(), positions = integer(), length = integer())
}
\arguments{
\item{values}{Vector of values}

\item{positions}{Vector of positions, must be integers and same length as
values.}

\item{length}{Length of resulting vector}
}
\value{
A sparse_vector object.
}
\description{
Create new sparse vector
}
\examples{
sparse_vector(1, 4, 10)
}
