% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_which_na.R
\name{sparse_which_na}
\alias{sparse_which_na}
\title{Which indices are Missing Values}
\usage{
sparse_which_na(x)
}
\arguments{
\item{x}{A sparse vector.}
}
\value{
A logical vector.
}
\description{
Which indices are Missing Values
}
\details{
This function, as with any of the other helper functions assumes that the
input \code{x} is a sparse numeric vector. This is done for performance reasons,
and it is thus the users responsibility to perform input checking.
}
\examples{
sparse_which_na(
  sparse_double(c(10, NA, 11), c(1, 50, 111), 1000)
)
}
\seealso{
\code{\link[=sparse_is_na]{sparse_is_na()}}
}
