% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevctrs-package.R
\docType{package}
\name{sparsevctrs-package}
\alias{sparsevctrs}
\alias{sparsevctrs-package}
\title{sparsevctrs: Sparse Vectors for Use in Data Frames}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides sparse vectors powered by ALTREP (Alternative Representations for R Objects) that behave like regular vectors, and can thus be used in data frames. Also provides tools to convert between sparse matrices and data frames with sparse columns and functions to interact with sparse vectors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/sparsevctrs}
  \item \url{https://r-lib.github.io/sparsevctrs/}
  \item Report bugs at \url{https://github.com/r-lib/sparsevctrs/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emil.hvitfeldt@posit.co} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Other contributors:
\itemize{
  \item Davis Vaughan \email{davis@posit.co} [contributor]
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
