% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{tibble_to_sparse}
\alias{tibble_to_sparse}
\title{Convert tibbles to sparse matrices}
\usage{
tibble_to_sparse(x)
}
\arguments{
\item{x}{a tibble}
}
\value{
a sparse matrix
}
\description{
Convert tibbles to sparse matrices
}
\examples{
mts <- mtcars
mts$new <- new_sparse_vector(c(1, 4, 8), c(8, 1, 20), nrow(mts))

tibble_to_sparse(mtcars)

tibble_to_sparse(mts)
}
\seealso{
\code{\link[=sparse_to_tibble]{sparse_to_tibble()}}
}
