% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{type-predicates}
\alias{type-predicates}
\alias{sparse_positions}
\alias{sparse_values}
\title{Sparse vector type checkers}
\usage{
sparse_positions(x)

sparse_values(x)
}
\arguments{
\item{x}{value to be checked.}
}
\description{
Sparse vector type checkers
}
\examples{
x_sparse <- sparse_double(c(pi, 5, 0.1), c(2, 5, 10), 10)
x_dense <- c(0, pi, 0, 0, 0.5, 0, 0, 0, 0, 0.1)

sparse_positions(x_sparse)
sparse_values(x_sparse)

sparse_positions(x_dense)
sparse_values(x_dense)
}
