% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{new_sparse_real}
\alias{new_sparse_real}
\title{Create sparse numeric vector}
\usage{
new_sparse_real(value, position, length)
}
\arguments{
\item{value}{Numeric vector, values of non-zero entries.}

\item{position}{integer vector, indices of non-zero entries.}

\item{length}{Integer, Length of vector.}
}
\description{
Create sparse numeric vector
}
\details{
For developers:

setting \code{options("sparsevctrs.verbose_materialize" = TRUE)} will print a
message each time a sparse vector has been forced to materialize.
}
