% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{new_sparse_real}
\alias{new_sparse_real}
\title{Create sparse numeric vector}
\usage{
new_sparse_real(x, len)
}
\arguments{
\item{x}{thing}

\item{len}{length of vector}
}
\description{
Create sparse numeric vector
}
