% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{sparse_double}
\alias{sparse_double}
\title{Create sparse double vector}
\usage{
sparse_double(values, positions, length)
}
\arguments{
\item{values}{Numeric vector, values of non-zero entries.}

\item{positions}{integer vector, indices of non-zero entries.}

\item{length}{Integer, Length of vector.}
}
\description{
Create sparse double vector
}
\details{
\code{values} and \code{positions} are expected to be the same length, and are allowed
to both have zero length.

Allowed values for \code{value} is double and integer values. integer values will
be coerced to doubles. Missing values such as \code{NA} and \code{NA_real_} are
allowed. Everything else is disallowed, This includes \code{Inf} and \code{NaN}.

\code{positions} should be integers or integer-like doubles. Everything else is
not allowed. Positions should furthermore be positive (\code{0} not allowed),
unique, and in increasing order. Lastly they should all be smaller that
\code{length}.

For developers:

setting \code{options("sparsevctrs.verbose_materialize" = TRUE)} will print a
message each time a sparse vector has been forced to materialize.
}
\examples{
sparse_double(numeric(), integer(), 10)

sparse_double(c(pi, 5, 0.1), c(2, 5, 10), 10)

str(
  sparse_double(c(pi, 5, 0.1), c(2, 5, 10), 1000000000)
)
}
