% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_dummy.R
\name{sparse_dummy}
\alias{sparse_dummy}
\title{Generate sparse dummy variables}
\usage{
sparse_dummy(x)
}
\arguments{
\item{x}{A factor.}
}
\value{
A list of sparse integer dummy variables.
}
\description{
Generate sparse dummy variables
}
\details{
Only factor variables can be used with \code{\link[=sparse_dummy]{sparse_dummy()}}. A call to
\code{as.factor()} would be required for any other type of data.
}
\examples{
x <- factor(c("a", "a", "b", "c", "d", "b"))

sparse_dummy(x)
}
