% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevctrs-package.R
\docType{package}
\name{sparsevctrs-package}
\alias{sparsevctrs}
\alias{sparsevctrs-package}
\title{sparsevctrs: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EmilHvitfeldt/sparsevctrs}
  \item \url{https://emilhvitfeldt.github.io/sparsevctrs/}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/sparsevctrs/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

}
\keyword{internal}
